PET-99 refactor : Image-Application S3 이미지 업로드 구현 코드 Image-Infrastructure 모듈로 이동
