+++ Domain-Module/Image-Module/Image-Infrastructure/src/main/java/com/pawith/imageinfrastructure/exception/FileUploadException.java
@@ -1,4 +1,4 @@
-package com.pawith.imageapplication.exception;
+package com.pawith.imageinfrastructure.exception;
 
 import com.pawith.commonmodule.exception.BusinessException;
 import com.pawith.commonmodule.exception.Error;
