+++ settings.gradle
@@ -26,4 +26,5 @@ include 'Domain-Module:Auth-Module:Auth-Domain'
 include 'Common-Module'
 include 'Domain-Module:Image-Module'
 include 'Domain-Module:Image-Module:Image-Application'
+include 'Domain-Module:Image-Module:Image-Infrastructure'
